Added helpful comments and fixed some code to ensure functionality continues after adding comments

+++ tech_blog/controllers/api/blogRoutes.js
@@ -1,49 +1,57 @@
-const express = require("express");
+const express = require('express');
 const router = express.Router();
-const { User, Blog, Comment } = require("../../models");
-const withAuth = require('../../util/auth');
+const { User, Blog, Comment } = require('../../models');
 
-// Get all blogs
-router.get("/", (req, res) => {
+// Get all blogs and their associated users and comments
+router.get('/', (req, res) => {
     Blog.findAll({ include: [User, Comment] })
         .then(dbBlogs => res.json(dbBlogs))
-        .catch(err => res.status(500).json({ msg: "an error occurred", err }));
+        .catch(err => res.status(500).json({ msg: 'An error occurred', err }));
 });
 
-// Get blog by id
-router.get("/:id", (req, res) => {
+// Get a specific blog by its ID, including the associated user and comments
+router.get('/:id', (req, res) => {
     Blog.findByPk(req.params.id, { include: [User, Comment] })
         .then(dbBlog => res.json(dbBlog))
-        .catch(err => res.status(500).json({ msg: "an error occurred", err }));
+        .catch(err => res.status(500).json({ msg: 'An error occurred', err }));
 });
 
-// Create new blog
-router.post("/", withAuth, (req, res) => {
+// Create a new blog post
+router.post('/', (req, res) => {
+    if (!req.session.user) {
+        return res.status(401).json({ msg: 'Please login!' });
+    }
     Blog.create({
         title: req.body.title,
         content: req.body.content,
         userId: req.session.user.id
     })
         .then(newBlog => res.json(newBlog))
-        .catch(err => res.status(500).json({ msg: "an error occurred", err }));
+        .catch(err => res.status(500).json({ msg: 'An error occurred', err }));
 });
 
-// Update blog
-router.put("/:id", withAuth, (req, res) => {
+// Update a blog post by its ID
+router.put('/:id', (req, res) => {
+    if (!req.session.user) {
+        return res.status(401).json({ msg: 'Please login!' });
+    }
     Blog.update(req.body, {
         where: { id: req.params.id }
     })
         .then(updatedBlog => res.json(updatedBlog))
-        .catch(err => res.status(500).json({ msg: "an error occurred", err }));
+        .catch(err => res.status(500).json({ msg: 'An error occurred', err }));
 });
 
-// Delete blog
-router.delete("/:id", withAuth, (req, res) => {
+// Delete a blog post by its ID
+router.delete('/:id', (req, res) => {
+    if (!req.session.user) {
+        return res.status(401).json({ msg: 'Please login!' });
+    }
     Blog.destroy({
         where: { id: req.params.id }
     })
         .then(delBlog => res.json(delBlog))
-        .catch(err => res.status(500).json({ msg: "an error occurred", err }));
+        .catch(err => res.status(500).json({ msg: 'An error occurred', err }));
 });
 
-module.exports = router;
+module.exports = router; // Export the blog routes
